Added composer

+++ admin/handlers/deleteUser.php
@@ -5,9 +5,11 @@
  * Date: 2017-11-29
  * Time: 1:41 PM
  */
-include "../../include.php";
-(function() {
-    $user = new User($_GET['id']);
-    $user->delete();
+require_once "../../include.php";
+(function () {
+    if ($GLOBALS['userLvl'] == Constants::LVL_ADMIN) {
+        $user = new User($_GET['id']);
+        $user->delete();
+    }
     Util::returnPrevPage();
 })();
